pyproject.toml
[project]
name = "crews"
version = "0.1.0"
description = "Set of crews for various tasks"
authors = [{ name = "herveGuigoz", email = "herve.guigoz@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.3.0",
    "youtube-transcript-api (>=1.2.3,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)"
]

[project.scripts]
youtube_transcript = "crews.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
